pyproject.toml
[project]
name = "price-identifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-kusto-data>=5.0.5",
    "dash>=3.2.0",
    "dash-bootstrap-components>=2.0.3",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pip-system-certs>=5.2",
    "plotly>=6.3.0",
    "pydantic>=2.11.7",
]
